+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
+### Unreleased
+
+#### Added
+
+* Add total coverage % to simplecov results on tests info extractor.
+
 ### v1.1.0
 
 #### Added